requirements.txt
dbt-bigquery==1.11.1
google-cloud-bigquery==3.25.0
openai==1.30.0
httpx==0.27.0
tqdm==4.66.0
python-dotenv==1.0.0
pandas==2.2.0
beautifulsoup4==4.12.0
langdetect
plotly
db-dtypes

